Extracted ta4j-csv module
